requirements.txt
aioredis==1.3.1
aiosmtplib==1.1.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
croniter==0.3.36
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-cors-headers==3.5.0
django-filter==2.4.0
django-rq==2.4.0
django-tinymce==3.2.0
djangorestframework==3.12.2
drf-spectacular==0.11.1
feedparser==6.0.2
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
http3==0.6.7
httptools==0.1.1
hyperframe==5.2.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
MarkupSafe==1.1.1
msgpack==1.0.1
natsort==7.1.0
nltk==3.5
oauthlib==3.1.0
packaging==20.7
pip-review==1.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rq==1.7.0
rq-scheduler==0.10.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tqdm==4.54.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.13.1
uvloop==0.14.0
websockets==8.1
whitenoise==5.2.0
zope.interface==5.2.0
